tester/requirements.txt

requests
numpy
python-Levenshtein
soundfile
